requirements.txt
absl-py==1.3.0
autopep8==2.0.0
beautifulsoup4==4.9.3
CacheControl==0.12.11
cachetools==5.2.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.1.3
colorzero==1.1
cryptography==3.4.2
dbus-python==1.2.16
distro==1.5.0
docopt==0.6.2
firebase-admin==6.0.1
Flask==2.2.2
Flask-Cors==3.0.10
flatbuffers==20181003210633
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
gpiozero==1.6.2
grpcio==1.50.0
grpcio-status==1.50.0
html5lib==1.1
httplib2==0.21.0
idna==2.10
ifaddr==0.1.7
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.6.3
MarkupSafe==2.1.1
msgpack==1.0.4
netifaces==0.10.9
notify2==0.3
numpy==1.23.5
opencv-python==4.6.0.66
paho-mqtt==1.6.1
picamera2==0.3.3
pidng==4.0.9
piexif==1.1.3
Pillow==8.1.2
proto-plus==1.22.1
protobuf==4.21.9
psutil==5.8.0
pulseaudio-dlna==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pybind11==2.10.1
pycairo==1.16.2
PyChromecast==7.7.1
pycodestyle==2.10.0
pycparser==2.21
pygame==2.1.2
PyGObject==3.38.0
PyJWT==2.6.0
pyparsing==3.0.9
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyroute2==0.5.14
PySDL2==0.9.7
python-apt==2.2.1
python-dotenv==0.21.0
python-prctl==1.7
pyttsx3==2.90
pytz==2022.6
requests==2.25.1
RPi.GPIO==0.7.0
rsa==4.9
RTIMULib==7.2.1
sense-hat==2.4.0
setproctitle==1.2.1
simplejpeg==1.6.4
six==1.16.0
sounddevice==0.4.5
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
tflite-support==0.4.3
toml==0.10.1
tomli==2.0.1
twilio==7.15.3
uritemplate==4.1.1
urllib3==1.26.5
v4l2-python3==0.3.1
webencodings==0.5.1
Werkzeug==2.2.2
zeroconf==0.26.1
zipp==3.10.0
